Another nice cleanup round, getting there .. eventually
