+++ app/base/User.js
@@ -1,0 +1,14 @@
+
+
+
+class User {
+
+    constructor( ) {
+
+    }
+
+    get identities() { return []; }
+
+}
+
+module.exports = User;
