+++ app/env.js
@@ -1,14 +1,13 @@
 /**
- * Canvas environment "bootstrap"
+ * Canvas environment variables
  */
 
 // Utils
 const path = require('path')
 const fs = require('fs')
 const os = require('os')
-const isElectron = require('is-electron')()
 const pkg = require('./package.json')
-const device = require('./utils/device')
+const isElectron = require('is-electron')()
 
 
 /**
@@ -23,7 +22,6 @@ const device = require('./utils/device')
  * |   ├── run
  * |   |   ├── canvas-ipc.sock
  * |   |   ├── canvas.pid
- * |   |   ├── canvas-user-role.sock
  * |   ├── tmp
  */
 
@@ -32,11 +30,14 @@ const APP_HOME = path.join(APP_ROOT, 'app')
 const APP_CONFIG = path.join(APP_ROOT, 'config')
 const APP_VAR = path.join(APP_ROOT, 'var')
 
+// Check for portable setup
+const isPortable = ! fs.existsSync(path.join(APP_ROOT, 'user', '.ignore'))
+
 
 /**
  * User directories
  *
- * APP_ROOT/user or ~/.canvas
+ * APP_ROOT/user or os.homedir/.canvas
  * ├── config
  * ├── cache
  * ├── data
@@ -53,40 +54,22 @@ const APP_VAR = path.join(APP_ROOT, 'var')
  */
 
 // User env
-const USER_HOME = getUserHome()
-const USER_CONFIG = path.join(USER_HOME, 'config')
-const USER_CACHE = path.join(USER_HOME, 'cache')
-const USER_DATA = path.join(USER_HOME, 'data')
-const USER_VAR = path.join(USER_HOME, 'var')
-
-
-// Initialize the global config module
-const Config = require('./utils/config')
-const config = Config({
-    userConfigDir: USER_CONFIG,
-    appConfigDir: APP_CONFIG,
-    versioning: false
-})
-
-// Initialize the logging module
-const Log = require('./utils/logger')
-const logger = new Log({
-    appName: pkg.name,
-    logLevel: process.env.LOG_LEVEL || 'debug',
-    logPath: path.join(USER_VAR, 'log')
-})
+const USER_HOME = process.env['CANVAS_USER_HOME'] || getUserHome()
+const USER_CACHE = process.env['CANVAS_USER_CACHE'] || path.join(USER_HOME, 'cache')
+const USER_CONFIG = process.env['CANVAS_USER_CONFIG'] || path.join(USER_HOME, 'config')
+const USER_DATA = process.env['CANVAS_USER_DATA'] || path.join(USER_HOME, 'data')
+const USER_VAR = process.env['CANVAS_USER_VAR'] || path.join(USER_HOME, 'var')
 
 
 const env = {
-    app: {
-        // App package info
+    APP: {
         name: (pkg.productName) ? pkg.productName : pkg.name,
         version: pkg.version,
         description: pkg.description,
         license: pkg.license,
         isElectron,
+        isPortable,
         paths: {
-            // App directories
             root: APP_ROOT,
             home: APP_HOME,
             config: APP_CONFIG,
@@ -94,9 +77,8 @@ const env = {
         }
     },
 
-    user: {
+    USER: {
         paths: {
-            // User directories
             home: USER_HOME,
             config: USER_CONFIG,
             cache: USER_CACHE,
@@ -105,17 +87,11 @@ const env = {
         }
     },
 
-    device,
-
     transport: {
         ipc: (process.platform === 'win32') ?
             path.join('\\\\?\\pipe', process.cwd(), pkg.name) :
             path.join(APP_VAR, 'run', `${pkg.name}-ipc.sock`)
-    },
-
-    config,
-    logger
-
+    }
 }
 
 
@@ -144,36 +120,15 @@ Object.assign(process.env, {
 
 });
 
-
 module.exports = env
-
 
 /**
  * Utils
  */
 
-// TODO: Rework, ugly
 function getUserHome() {
-
-    // If CANVAS_USER_HOME was supplied explicitly, return it.
-    if (process.env.CANVAS_USER_HOME) {
-        return process.env.CANVAS_USER_HOME;
-    }
-
-    // If CANVAS_PORTABLE was set, return the portable user dir.
-    if (process.env.CANVAS_PORTABLE) {
-        return path.join(APP_ROOT, 'user');
-    }
-
-    // Check for portable setup and return user directory if portable mode is disabled.
-    if (!fs.existsSync(path.join(APP_ROOT, 'user', '.ignore'))) {
-        return path.join(APP_ROOT, 'user');
-    }
-
-    // Default: Fallback to the local OS home directory.
-    return path.join(os.homedir(), ".canvas");
+    return (isPortable) ? path.join(APP_ROOT, 'user') : path.join(os.homedir(), ".canvas");
 }
-
 
 function checkObjectAgainstSchema(obj, schema) {
