+++ app/package.json
@@ -20,11 +20,13 @@
     "conf": "^10.2.0",
     "date-fns": "^2.30.0",
     "debug": "^4.3.4",
+    "detect-libc": "^2.0.2",
     "eventemitter2": "^6.4.9",
     "express": "^4.18.2",
     "file-type": "=18.5.0",
     "forever-monitor": "^1.5.2",
     "fs-extra": "^11.1.1",
+    "ip": "^2.0.0",
     "is-electron": "^2.2.2",
     "lmdb": "^2.8.1",
     "magic-bytes.js": "^1.0.14",
