--- app/utils/device/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "canvas-utils-device",
-  "version": "1.0.0",
-  "main": "index.js",
-  "author": "idnc_sk",
-  "license": "AGPL-3.0-or-later",
-  "private": true,
-  "dependencies": {
-    "detect-libc": "^2.0.1",
-    "ip": "^1.1.8",
-    "node-machine-id": "^1.1.12"
-  }
-}
