--- bin/canvas-installer.sh
@@ -1,27 +1,0 @@
-#!/bin/bash
-
-
-# Download runtimes
-
-# Download models
-
-
-# Regenerate API keys and update config
-
-
-# Update system PATH
-
-# Update system aliases
-
-# Install browser extension
-
-
-# Toggle portable mode
-
-# Toggle startOnBoot
-
-# Toggle startAsSwayOverlay
-
-# Toggle startAsPicomOverlay
-
-# Toggle startDesktopMode
