+++ bin/install.sh
@@ -15,4 +15,31 @@ CANVAS_ELECTRON_VERSION=""
 
 # Check dependencies
 
+#
 # Install
+
+
+# Download runtimes
+
+# Download models
+
+
+# Regenerate API keys and update config
+
+
+# Update system PATH
+
+# Update system aliases
+
+# Install browser extension
+
+
+# Toggle portable mode
+
+# Toggle startOnBoot
+
+# Toggle startAsSwayOverlay
+
+# Toggle startAsPicomOverlay
+
+# Toggle startDesktopMode
