--- bin/ws
@@ -1,8 +1,0 @@
-#!/usr/bin/env node
-
-// Assuming node is installed and within PATH
-// TODO: Change to runtime dir
-// Enable the below if --use-system-nodejs is supplied
-const path  = require('path')
-const app = path.join(path.dirname(__dirname), 'app', 'ui', 'cli', 'cli.js')
-require(app)
